ai_models/requirements.txt
mediapipe==0.10.8
opencv-python==4.8.1.78
numpy==1.24.3
scikit-learn==1.3.2
pytesseract==0.3.10
gtts==2.5.1
pydub==0.25.1
pyaudio==0.2.14
speechrecognition==3.10.0
Pillow==10.1.0

